implementando as regras de neg'ocio

+++ src/main/java/br/com/blz/testjava/product/ProductServiceImpl.java
@@ -1,10 +1,13 @@
 package br.com.blz.testjava.product;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.blz.testjava.error.NoProductResultException;
 import br.com.blz.testjava.error.ProductSavedException;
+import br.com.blz.testjava.warehouse.Warehouse;
 
 @Service
 public class ProductServiceImpl implements ProductService {
@@ -18,12 +21,20 @@ public class ProductServiceImpl implements ProductService {
 		if(p != null) {
 			throw new ProductSavedException("There is a product with sku " + product.getSku() + " saved.");
 		}
-		return repository.save(product);
+		product = repository.save(product);
+		product = setInventoryQuantity(product);	
+		product = isMarketable(product);
+		return product;
 	}
 
 	@Override
 	public Product getBySku(Long sku) {
-		return repository.findBySku(sku);
+		Product product = repository.findBySku(sku);
+		if(product != null) {
+			product = setInventoryQuantity(product);	
+			product = isMarketable(product);
+		}
+		return product;
 	}
 
 	@Override
@@ -32,13 +43,45 @@ public class ProductServiceImpl implements ProductService {
 		if(p == null) {
 			throw new NoProductResultException("No product with sku " + product.getSku());
 		}
-		return repository.update(product);
+		p = repository.update(product);
+		p = setInventoryQuantity(product);	
+		p = isMarketable(product);
+		return p;
 	}
 
 	@Override
-	public void deleteBySku(Long sku) {
+	public void deleteBySku(Long sku) throws NoProductResultException {
+		Product p = repository.findBySku(sku); 
+		if(p == null) {
+			throw new NoProductResultException("No product with sku " + sku);
+		}
 		repository.delete(sku);
 
+	}
+	
+	private Product setInventoryQuantity(Product product) {
+		Integer quantity = 0;
+		List<Warehouse> warehouses = product.getInventory().getWarehouses();
+		
+		for (Warehouse warehouse : warehouses) {
+			quantity += warehouse.getQuantity();
+		}
+		
+		product.getInventory().setQuantity(quantity);
+		
+		return product;
+	}
+	
+	private Product isMarketable(Product product) {
+		Integer inventoryQuantity = product.getInventory().getQuantity();
+		boolean isMarketable = false;
+		if(inventoryQuantity > 0) {
+			isMarketable = true;
+		}
+		
+		product.setIsMarketable(isMarketable);
+		
+		return product;
 	}
 
 }
